refactor(docs): simplify .md routes with single catch-all
Replace 4 separate routes + 16 rewrites with:
- 1 unified route that handles all sources
- 2 rewrite rules: /:path*.md and /:path*.mdx
