--- docs/app/llms.mdx/reference/[[...slug]]/route.ts
@@ -1,24 +1,0 @@
-import { referenceSource } from '@/lib/source';
-import { getLLMText } from '@/lib/source';
-import { notFound } from 'next/navigation';
-
-export const revalidate = false;
-
-export async function GET(
-  _req: Request,
-  { params }: { params: Promise<{ slug?: string[] }> }
-) {
-  const { slug } = await params;
-  const page = referenceSource.getPage(slug);
-  if (!page) notFound();
-
-  return new Response(await getLLMText(page), {
-    headers: {
-      'Content-Type': 'text/markdown; charset=utf-8',
-    },
-  });
-}
-
-export function generateStaticParams() {
-  return referenceSource.generateParams();
-}
